requirements.txt
streamlit>=1.29.0
openai>=1.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
